+++ mne/export/__init__.py
@@ -1,1 +1,2 @@
-from ._export import export_raw, export_epochs
+from ._export import export_raw, export_epochs, export_evokeds
+from ._egimff import export_evokeds_mff
