+++ mne/export/tests/test_export.py
@@ -11,10 +11,18 @@ import pytest
 import numpy as np
 from numpy.testing import assert_allclose, assert_array_equal
 
-from mne import read_epochs_eeglab, Epochs
-from mne.tests.test_epochs import _get_data
+from mne import read_epochs_eeglab, Epochs, read_evokeds, read_evokeds_mff
+from mne.datasets import testing
+from mne.export import export_evokeds, export_evokeds_mff
 from mne.io import read_raw_fif, read_raw_eeglab
-from mne.utils import _check_eeglabio_installed
+from mne.utils import _check_eeglabio_installed, requires_version, object_diff
+from mne.tests.test_epochs import _get_data
+
+base_dir = op.join(op.dirname(__file__), '..', '..', 'io', 'tests', 'data')
+fname_evoked = op.join(base_dir, 'test-ave.fif')
+
+data_path = testing.data_path(download=False)
+egi_evoked_fname = op.join(data_path, 'EGI', 'test_egi_evoked.mff')
 
 
 @pytest.mark.skipif(not _check_eeglabio_installed(strict=False),
@@ -62,3 +70,75 @@ def test_export_epochs_eeglab(tmpdir, preload):
     assert epochs.event_id.keys() == epochs_read.event_id.keys()  # just keys
     assert_allclose(epochs.times, epochs_read.times)
     assert_allclose(epochs.get_data(), epochs_read.get_data())
+
+
+@requires_version('mffpy', '0.5.7')
+@testing.requires_testing_data
+@pytest.mark.parametrize('fmt', ('auto', 'mff'))
+@pytest.mark.parametrize('do_history', (True, False))
+def test_export_evokeds_to_mff(tmpdir, fmt, do_history):
+    """Test exporting evoked dataset to MFF."""
+    evoked = read_evokeds_mff(egi_evoked_fname)
+    export_fname = op.join(str(tmpdir), 'evoked.mff')
+    history = [
+        {
+            'name': 'Test Segmentation',
+            'method': 'Segmentation',
+            'settings': ['Setting 1', 'Setting 2'],
+            'results': ['Result 1', 'Result 2']
+        },
+        {
+            'name': 'Test Averaging',
+            'method': 'Averaging',
+            'settings': ['Setting 1', 'Setting 2'],
+            'results': ['Result 1', 'Result 2']
+        }
+    ]
+    if do_history:
+        export_evokeds_mff(export_fname, evoked, history=history)
+    else:
+        export_evokeds(export_fname, evoked)
+    # Drop non-EEG channels
+    evoked = [ave.drop_channels(['ECG', 'EMG']) for ave in evoked]
+    evoked_exported = read_evokeds_mff(export_fname)
+    assert len(evoked) == len(evoked_exported)
+    for ave, ave_exported in zip(evoked, evoked_exported):
+        # Compare infos
+        assert object_diff(ave_exported.info, ave.info) == ''
+        # Compare data
+        assert_allclose(ave_exported.data, ave.data)
+        # Compare properties
+        assert ave_exported.nave == ave.nave
+        assert ave_exported.kind == ave.kind
+        assert ave_exported.comment == ave.comment
+        assert_allclose(ave_exported.times, ave.times)
+
+
+@requires_version('mffpy', '0.5.7')
+@testing.requires_testing_data
+def test_export_to_mff_no_device():
+    """Test no device type throws ValueError."""
+    evoked = read_evokeds_mff(egi_evoked_fname, condition='Category 1')
+    evoked.info['device_info'] = None
+    with pytest.raises(ValueError, match='No device type.'):
+        export_evokeds('output.mff', evoked)
+
+
+@requires_version('mffpy', '0.5.7')
+def test_export_to_mff_incompatible_sfreq():
+    """Test non-whole number sampling frequency throws ValueError."""
+    evoked = read_evokeds(fname_evoked)
+    with pytest.raises(ValueError, match=f'sfreq: {evoked[0].info["sfreq"]}'):
+        export_evokeds('output.mff', evoked)
+
+
+@pytest.mark.parametrize('fmt,ext', [
+    ('EEGLAB', 'set'),
+    ('EDF', 'edf'),
+    ('BrainVision', 'eeg')
+])
+def test_export_evokeds_unsupported_format(fmt, ext):
+    """Test exporting evoked dataset to non-supported formats."""
+    evoked = read_evokeds(fname_evoked)
+    with pytest.raises(NotImplementedError, match=f'Export to {fmt} not imp'):
+        export_evokeds(f'output.{ext}', evoked)
