+++ mne/io/egi/__init__.py
@@ -3,4 +3,4 @@
 # Author: Denis A. Engemann <denis.engemann@gmail.com>
 
 from .egi import read_raw_egi
-from .egimff import read_evokeds_mff, export_evokeds_to_mff
+from .egimff import read_evokeds_mff
